Added content from Kent C Dodds 'React Fundamentals' course
So in here, we're setting up the stage by copying over the resources made available by Kent. For the rest of this branch, we'll be working on duplicating HTML files and completing the course work with Feliz & Fable. We'll reuse the CSS files through using TypedCssClasses to get typed CSS types directly accessible in our F# source code.